BackupUserProfile
This script performs a backup of user profile to a remote server, log
the information on log file and delete the profile in case of
specification. Before delete, it compares if source and destination have
the same size and delete only if it is equal. If you don't want that,
change the script. The SourceUserFile is a text file with usernames of
profile to backup and, eventually, delete. Use one user profile per
line.